Add in config layer for stageconfig + minor review related changes
