+++ src/test/java/com/google/cloud/tools/maven/util/SingleYamlFlexibleDeployTestHelper.java
@@ -40,6 +40,8 @@ public class SingleYamlFlexibleDeployTestHelper<M extends AbstractDeployMojo>
   public void before() throws IOException {
     mojo.setStagingDirectory(temporaryFolder.newFolder("staging"));
     mojo.setSourceDirectory(temporaryFolder.newFolder("source"));
+    mojo.setProject("some-project");
+    mojo.setVersion("some-version");
     MockitoAnnotations.initMocks(this);
 
     when(mavenProject.getProperties()).thenReturn(new Properties());
